Split upscale pricing preview hook
