+++ tests/upscale-workspace-architecture.test.ts
@@ -9,6 +9,7 @@ const copyPath = join(root, 'frontend/src/components/tools/upscale/_lib/upscale-
 const typesPath = join(root, 'frontend/src/components/tools/upscale/_lib/upscale-workspace-types.ts');
 const helpersPath = join(root, 'frontend/src/components/tools/upscale/_lib/upscale-workspace-helpers.ts');
 const libraryHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks/useUpscaleLibraryAssets.ts');
+const pricingHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks/useUpscalePricingPreview.ts');
 
 const workspaceSource = readFileSync(workspacePath, 'utf8');
 
@@ -17,11 +18,13 @@ test('upscale workspace delegates copy, local types, and helper logic', () => {
   assert.ok(existsSync(typesPath), 'upscale workspace local contracts should live in a colocated type module');
   assert.ok(existsSync(helpersPath), 'upscale workspace helper logic should live in a colocated helper module');
   assert.ok(existsSync(libraryHookPath), 'upscale library asset loading should live in a colocated hook');
+  assert.ok(existsSync(pricingHookPath), 'upscale pricing preview should live in a colocated hook');
 
   assert.match(workspaceSource, /from '\.\/upscale\/_lib\/upscale-workspace-copy'/, 'workspace should import upscale copy');
   assert.match(workspaceSource, /from '\.\/upscale\/_lib\/upscale-workspace-types'/, 'workspace should import upscale local types');
   assert.match(workspaceSource, /from '\.\/upscale\/_lib\/upscale-workspace-helpers'/, 'workspace should import upscale helpers');
   assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscaleLibraryAssets'/, 'workspace should import library hook');
+  assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscalePricingPreview'/, 'workspace should import pricing hook');
 });
 
 test('upscale workspace does not regain extracted ownership', () => {
@@ -33,9 +36,12 @@ test('upscale workspace does not regain extracted ownership', () => {
   assert.doesNotMatch(workspaceSource, /UserAssetsResponse/, 'library asset response parsing belongs in useUpscaleLibraryAssets');
   assert.doesNotMatch(workspaceSource, /JobsLibraryResponse/, 'generated video library merging belongs in useUpscaleLibraryAssets');
   assert.doesNotMatch(workspaceSource, /buildLibraryCacheKey/, 'library cache key orchestration belongs in useUpscaleLibraryAssets');
+  assert.doesNotMatch(workspaceSource, /buildUpscalePricingPreview/, 'pricing preview orchestration belongs in useUpscalePricingPreview');
+  assert.doesNotMatch(workspaceSource, /readVideoPricingMetadata/, 'video pricing metadata loading belongs in useUpscalePricingPreview');
+  assert.doesNotMatch(workspaceSource, /BillingProductResponse/, 'billing product parsing belongs in useUpscalePricingPreview');
 
   const lineCount = workspaceSource.split('\n').length;
-  assert.ok(lineCount <= 1260, `UpscaleWorkspace should stay below 1260 lines after library hook extraction, got ${lineCount}`);
+  assert.ok(lineCount <= 1195, `UpscaleWorkspace should stay below 1195 lines after pricing hook extraction, got ${lineCount}`);
 });
 
 test('upscale helper modules expose the expected workspace contract', () => {
@@ -43,6 +49,7 @@ test('upscale helper modules expose the expected workspace contract', () => {
   const typesSource = readFileSync(typesPath, 'utf8');
   const helpersSource = readFileSync(helpersPath, 'utf8');
   const libraryHookSource = readFileSync(libraryHookPath, 'utf8');
+  const pricingHookSource = readFileSync(pricingHookPath, 'utf8');
 
   assert.match(copySource, /export const DEFAULT_UPSCALE_COPY =/, 'copy module should export default copy');
 
@@ -74,4 +81,8 @@ test('upscale helper modules expose the expected workspace contract', () => {
   assert.match(libraryHookSource, /buildLibraryCacheKey/, 'library hook should own cache key usage');
   assert.match(libraryHookSource, /UserAssetsResponse/, 'library hook should parse saved asset responses');
   assert.match(libraryHookSource, /JobsLibraryResponse/, 'library hook should merge generated video jobs');
+  assert.match(pricingHookSource, /export function useUpscalePricingPreview/, 'pricing hook should be exported');
+  assert.match(pricingHookSource, /buildUpscalePricingPreview/, 'pricing hook should build the preview');
+  assert.match(pricingHookSource, /readVideoPricingMetadata/, 'pricing hook should load video metadata');
+  assert.match(pricingHookSource, /BillingProductResponse/, 'pricing hook should parse billing products');
 });
